pyproject.toml
[project]
name = "filmlayout"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "pyside6",
    "pillow",
]
